start refactoring tests
